create a bucket for MSSM to transfer files
This is related to commit 335107f

Our collaborators at MSSM do not want to create an AWS account for
us to setup cross account access for them to access one of our buckets
so they can transfer files to Sage.

After further discussion with MSSM we've decided to provide MSSM
with an Sage IAM account user access key with access to the bucket.