[IT-489] Option to open a port to incoming traffic
Users may want to run an application on the EC2 instance and open
a port to it so other can access the app from a client.  This
allows opening a port for incoming traffic to the EC2.

It is not a required parameter, if nothing is specified no ports
will be open.