Fix ubuntu deployment (#125)
This is failing on travis only, it works when deployed locally.  The
problem seems to be that it's using the aws linux template[1] instead
of the ubuntu template.  My guess is that travis is over aggressively
cacheing files therefore not replacing the aws linux template.

We attempt to fix by downloading to a different file.